// src/enums/modeling.enum.ts
export enum ModelHeightEnum {
    FT = 'ft',
    CM = 'cm',
    IN ='in'
}


export enum ModelWeightEnum {
    KG = 'kg',
    LBS = 'lbs'
}

export enum ModelBodySizeEnum {
    IN = 'in',
    CM = 'cm'
}


export enum YesNoEnum {
    YES = 'Yes',
    NO = 'No'
}


export enum OrderByEnum {
    ASCENDING = "ascending",
    DESCENDING = "descending"
}


export enum ModelStatsEnum {
    PROFILE = "profile",
    PORTFOLIO = "portfolio",
    BASIC = "basic",
    ADVANCE = "advance",
    PHOTOS = "photos",
}


export enum DefaultUnit {
    IMPERIAL = "imperial ",
    METRIC = "metric"
}


export enum FilterSearchType {
    BASIC = "basic",
    ADVANCE = "advance"
}


export enum ChatMessageType {
    MESSAGE = "message",
    ASSET = "asset"
}


export enum OfflineDaysEnumType {
    ONE = 1,
    THREE = 3,
    SEVEN = 7,
    FOURTEEN = 14,
    THIRTY = 30,

}


export enum ApprovalStatus {
    APPROVED = "approved",
    ACCEPTED = "accepted",
    REJECTED = "rejected",
    PENDING = "pending"

}


export enum AssetType {
    IMAGE = "image",
    VIDEO = "video"
}


export enum BuyPortfolioType {
    SINGLE_USE = "single_use",
    UNLIMITED_USE_COUNTRY = "unlimited_use_country",
    UNLIMITED_USE_GLOBAL = "unlimited_use_global"

}